add policy for admin, role & user

+++ app/Http/Controllers/Admins/RoleController.php
@@ -5,36 +5,49 @@ namespace App\Http\Controllers\Admins;
 use App\Http\Controllers\Controller;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class RoleController extends Controller
 {
     public function index()
     {
-        return Inertia::render('Admin/Roles/Index',[
-            'roles'=>Role::all()
-        ]);
+        if(Gate::allows('accessRoles')) {
+            return Inertia::render('Admin/Roles/Index',[
+                'roles'=>Role::all()
+            ]);
+        }
+        return back();
     }
 
     public function create()
     {
-        return Inertia::render('Admin/Roles/Create');
+        if(Gate::allows('manageRoles')) {
+            return Inertia::render('Admin/Roles/Create');
+        }
+        return back();
     }
 
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required'
-        ]);
-        Role::create($request->only(['name']));
-        return redirect(route('admin.roles.index'))->withSuccess('Role created successfully!');
+        if(Gate::allows('manageRoles')) {
+            $request->validate([
+                'name' => 'required'
+            ]);
+            Role::create($request->only(['name']));
+            return redirect(route('admin.roles.index'))->withSuccess('Role created successfully!');
+        }
+        return back();
     }
 
     public function show(Role $role)
     {
-        return Inertia::render('Admin/Roles/Show', [
-            'role' => $role,
-        ]);
+        if(Gate::allows('accessRoles')) {
+            return Inertia::render('Admin/Roles/Show', [
+                'role' => $role,
+            ]);
+        }
+        return back();
     }
 
     public function edit(Role $role)
@@ -44,16 +57,22 @@ class RoleController extends Controller
 
     public function update(Request $request, Role $role)
     {
-        $request->validate([
-            'name' => 'required'
-        ]);
-        $role->update($request->only(['name']));
-        return redirect(route('admin.roles.index'))->withSuccess('Role updated successfully!');
+        if(Gate::allows('manageRoles')) {
+            $request->validate([
+                'name' => 'required'
+            ]);
+            $role->update($request->only(['name']));
+            return redirect(route('admin.roles.index'))->withSuccess('Role updated successfully!');
+        }
+        return back();
     }
 
     public function destroy(Role $role)
     {
-        $role->delete();
-        return redirect(route('admin.roles.index'))->withSuccess('Role deleted successfully!');
+        if(Gate::allows('manageRoles')) {
+            $role->delete();
+            return redirect(route('admin.roles.index'))->withSuccess('Role deleted successfully!');
+        }
+        return back();
     }
 }
